refactor(al_loop): separated mlip benchmark logic into modules
